Add /cars endpoint test

+++ app/controllers/cars_controller.rb
@@ -10,7 +10,11 @@ class CarsController < ApplicationController
 
   # GET /cars/1
   def show
-    render json: { status: { code: 200, message: 'Fetch car detail successfully.' }, data: @car }, status: :ok
+    if @car
+      render json: { status: { code: 200, message: 'Fetch car detail successfully.' }, data: @car }, status: :ok
+    else
+      render json: { status: 404, message: 'Car not found' }, status: :not_found
+    end
   end
 
   # POST /cars
@@ -26,10 +30,14 @@ class CarsController < ApplicationController
 
   # DELETE /cars/1
   def destroy
-    if @car.destroy
-      render json: { status: { code: 200, message: 'Delete successfully.' } }, status: :ok
+    if @car
+      if @car.destroy
+        render json: { status: { code: 200, message: 'Delete successfully.' } }, status: :ok
+      else
+        render json: { status: 400, message: "Couldn't delete car" }, status: :bad_request
+      end
     else
-      render json: { status: 400, message: "Couldn't delete car" }, status: :bad_request
+      render json: { status: 404, message: 'Car not found' }, status: :not_found
     end
   end
 
@@ -38,6 +46,8 @@ class CarsController < ApplicationController
   # Use callbacks to share common setup or constraints between actions.
   def set_car
     @car = Car.find(params[:id])
+  rescue ActiveRecord::RecordNotFound
+    @car = nil
   end
 
   # Only allow a list of trusted parameters through.
